requirements.txt
pymupdf
faiss-cpu
openai
fastapi
uvicorn[standard]
python-multipart
langdetect
tiktoken
langgraph
langchain
sentence-transformers
watchdog